requirements.txt
SQLAlchemy~=2.0.40
uvicorn~=0.34.0
fastapi~=0.115.12
starlette~=0.46.1
python-dateutil~=2.9.0.post0
pydantic~=2.11.3
alembic~=1.15.2